requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
redis
fastapi-mail
pydantic-settings
python-multipart
email-validator
httpx
pytest
pytest-asyncio
typer
rich
# Auth dependencies
bcrypt
python-jose[cryptography]
slowapi
